Add endpoint to get release (#17)
