+++ requirements.txt
@@ -5,3 +5,5 @@ requests
 netifaces
 Flask-SQLAlchemy
 SQLAlchemy~=1.3.19
+gunicorn
+gevent
